Remove dwi, anatomical, and fieldmap workflows.
