--- qsirecon/workflows/anatomical/__init__.py
@@ -1,1 +1,0 @@
-from .volume import init_anat_preproc_wf, init_synthseg_wf, init_synthstrip_wf
