--- qsirecon/workflows/dwi/__init__.py
@@ -1,11 +1,0 @@
-"""
-Pre-processing q-space Image workflows
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-.. automodule:: qsirecon.workflows.dwi.base
-.. automodule:: qsirecon.workflows.dwi.util
-.. automodule:: qsirecon.workflows.dwi.hmc
-.. automodule:: qsirecon.workflows.dwi.fsl
-.. automodule:: qsirecon.workflows.dwi.registration
-.. automodule:: qsirecon.workflows.dwi.resampling
-.. automodule:: qsirecon.workflows.dwi.confounds
-"""
